requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
matplotlib==3.8.2
python-pptx==0.6.23
numpy==1.26.2
scipy==1.11.4
